compile without warnings

+++ classes/elements.cpp
@@ -6,22 +6,22 @@
 
 using namespace std;
 
-ElementContainer::ElementContainer(int initialCapacity)
+ElementContainer::ElementContainer(size_t initialCapacity)
     : capacity(initialCapacity), size(0) {
     elements = new Element*[capacity];
 }
 
 ElementContainer::~ElementContainer() {
-    for (int i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
         delete elements[i];
     }
     delete[] elements;
 }
 
 void ElementContainer::resize() {
-    int newCapacity = capacity * 2;
+    size_t newCapacity = capacity * 2;
     Element** newElements = new Element*[newCapacity];
-    for (int i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
         newElements[i] = elements[i];
     }
     delete[] elements;
@@ -36,19 +36,19 @@ void ElementContainer::addElement(const Element& element) {
     elements[size++] = new Element(element);
 }
 
-void ElementContainer::removeElement(int index) {
+void ElementContainer::removeElement(size_t index) {
     if (index >= size) {
         cout << "Выход за пределы индекса" << endl;
         return;
     }
     delete elements[index];
-    for (int i = index; i < size - 1; ++i) {
+    for (size_t i = index; i < size - 1; ++i) {
         elements[i] = elements[i + 1];
     }
     --size;
 }
 
-void ElementContainer::editElement(int index, const Element& updatedElement) {
+void ElementContainer::editElement(size_t index, const Element& updatedElement) {
     if (index >= size) {
         cout << "Выход за пределы индекса" << endl;
         return;
@@ -57,7 +57,7 @@ void ElementContainer::editElement(int index, const Element& updatedElement) {
 }
 
 void ElementContainer::findElementByMass(const int& mass) const {
-    for (int i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
         if (elements[i]->getMass() == mass) {
             elements[i]->shortDisplay();
         }
@@ -68,7 +68,7 @@ void ElementContainer::writeToJSON(const string& filename) const {
     json jsonFile;
     jsonFile["count"] = size;
 
-    for (int i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
         jsonFile["elements"].push_back(elements[i]->getJSON());
     }
 
@@ -92,8 +92,8 @@ void ElementContainer::readFromJSON(const string& filename) {
     json jsonFile;
     inFile >> jsonFile;
 
-    int newSize = jsonFile["count"];
-    for (int i = 0; i < size; ++i) {
+    size_t newSize = jsonFile["count"];
+    for (size_t i = 0; i < size; ++i) {
         delete elements[i];
     }
     delete[] elements;
@@ -110,13 +110,13 @@ void ElementContainer::readFromJSON(const string& filename) {
 }
 
 void ElementContainer::displayElements() const {
-    for (int i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
         elements[i]->display();
     }
 }
 
 void ElementContainer::displayMetals() const {
-    for (int i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
         if (elements[i]->isType("metal")) {
            elements[i]-> shortDisplay();
         }
